+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.7.0 (WIP)
 * add HMAC tool.
+* add Lines merging tool: add the lines of B to A if they're new.
+* add Lines subtract tool: remove the lines of B from A.
 
 ## 1.6.0 (2023/01/07)
 * fix some bugs with Timestamp tool.